chore(front): rename speaker-quota file
