+++ client/src/types/global-models.ts
